Fix retarded casting in castLong.

+++ src/main/java/me/taylorkelly/bigbrother/BetterConfig.java
@@ -24,7 +24,7 @@ public class BetterConfig extends Configuration {
         } else if (o instanceof Byte) {
             return (long) (Byte) o;
         } else if (o instanceof Integer) {
-            return (Long) o;
+            return Long.valueOf((Integer)o);
         } else if (o instanceof Double) {
             return (long) (double) (Double) o;
         } else if (o instanceof Float) {
